model-api/requirements.txt
Flask==3.1.1
Accelerate==1.8.1 
Torch==2.7.1
Transformers==4.39.3
Flask-Cors==5.0.0
MarkupSafe==3.0.2 
diskcache==5.6.3 
jinja2==3.1.6 
llama-cpp-python==0.3.14 
typing-extensions==4.14.1


# command to download the model in gguf format: huggingface-cli download microsoft/Phi-3-mini-4k-instruct-gguf \ Phi-3-mini-4k-instruct-q4.gguf \ --local-dir .